Add clustering part to heatmap to generate overall clustering network
